it-career-recommender/backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pandas>=2.0
numpy>=1.26
scikit-learn>=1.5.2
pdfplumber==0.11.4
docx2txt==0.9
sentence-transformers==3.0.1
spacy==3.8.7
# Optional: small English model for basic NLP; run: python -m spacy download en_core_web_sm
